Refactor : SettingsBarEntryProvider - Factory Method Pattern
Factory Method Pattern을
적용시키기 위해 추상 클래스인
SettingsBarEntryProviderVer1 구현
